+++ src/USI_I2C_Master.h
@@ -105,8 +105,8 @@
 
 void    i2c_init(void);
 uint8_t i2c_get_state_info(void);
-unsigned int i2c_start(uint8_t);
-unsigned int i2c_rep_start(uint8_t);
+uint8_t i2c_start(uint8_t);
+uint8_t i2c_rep_start(uint8_t);
 uint8_t i2c_stop(void);
 uint8_t i2c_write(uint8_t);
 uint8_t i2c_read(uint8_t);
